smbios: Extract SMBIOS table building code to separate function
Move the code that builds the SMBIOS tables to a separate
smbios_build_tables() function, to keep it isolated from the code
that initializes the SMBIOS entry point.

Thew new function will still take a smbios_21_entry_point
argument to make code review easier, but this will be changed by
the next commits.

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>