sercon: Fix keycodes for F11 and F12
Fix what happens when F11/F12 is sent via serial port. The existing code
already correctly detects the byte sequence that is sent on the serial
connection, however it translates it to the incorrect keycode. The
termseq table must map to *DOS scan codes*.

The incorrect value matches *PS2 set1 scan codes* for F11 and F12, so
fix the implementation mistake.

The patch is tested by Andrei on real hardware.

Signed-off-by: Michał Żygowski <michal.zygowski@3mdeb.com>
Cc: Andrei Purdea <seabios@purdea.ro>
[Add and message description by Andrei Purdea <seabios@purdea.ro>]
Link: https://mail.coreboot.org/archives/list/seabios@seabios.org/thread/V2KN2YKCL5J2JL4IMJKVTVE6YALIIC73/
Link: https://github.com/Dasharo/SeaBIOS/commit/0f4530859f11b3235b7d6163388f19577bdc6e1e
Fixes: d6728f30 ("add serial console support")
Signed-off-by: Paul Menzel <pmenzel@molgen.mpg.de>